requirements.txt
scikit-learn
torch
torchvision
pillow
numpy
seaborn
matplotlib
opencv-python